requirements.txt
   gradio==5.6.0
   autogen==0.4
   python-dotenv==1.0.0
   # Add other dependencies as needed